yet another failed attempt; TypeProviders cannot act on Types, only primitive literals
there might be a way to structure the projects so that the assembly
containing the type provider actually references the targeted types,
therefore allowing it to reflect on the types by name, but that would
require dividing the PeerNetwork assembly into 3 parts: one for the TP,
one for the protocol message types (to be referenced by the TP), and one
more to consume the TP and anything else.  That just seems like too much.