first stab at some client stuff; realized I want to refactor a bit
there are going to be 3 libraries here: Client, Server, and Protocol.  It occurs to me that
PeerNetwork is a vague naem, where as Protocol means exactly what I'm trying to say,
so I'm going to rename PeerNetwork to Protocol.  I also don't really like the
MessagePayload nomenclature, so I'll probably refactor that a bit also.